Create files for each class
- new files in lib
